% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum.R
\name{maximum}
\alias{maximum}
\title{Find maximum and minimum}
\usage{
maximum(x)
}
\arguments{
\item{x}{vector, a vector of numbers of length N}
}
\value{
return the maximum and minimum values
}
\description{
Find maximum and minimum
}
\details{
This function helps to find the maximum and minimum values of the variable
}
\examples{
maximum(rnorm(10,1,2))

}
