% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum.R
\name{minimum}
\alias{minimum}
\title{Find maximum and minimum}
\usage{
minimum(x)
}
\arguments{
\item{x}{vector, a vector of numbers of length N}
}
\value{
return the maximum and minimum values
}
\description{
Find maximum and minimum
}
\details{
This function helps to find the maximum and minimum values of the variable
}
\examples{
minimum(rnorm(10,2,1))

}
